fix(#884): move InMemoryCacheStore to contracts/ to fix import-linter boundary violation
The __getattr__ lazy re-export created a contracts -> bricks import chain
that import-linter detected. Moving the class directly into contracts/
(alongside CacheStoreABC and NullCacheStore) fixes the tier architecture
violation while keeping bricks/cache/inmemory.py as a thin re-export.
